+++ src/pages/_app.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect } from 'react';
 import toast from 'react-hot-toast';
-import { useUpdateAtom } from 'jotai/utils';
 import { useRouter } from 'next/router';
 import { AtomDebug, Devtools } from '@features/devtools';
 import { AppConfig } from '@components/app-config';
@@ -13,33 +12,25 @@ import 'tippy.js/dist/tippy.css';
 import 'modern-normalize/modern-normalize.css';
 import { getNetworkMode } from '@common/api/network';
 import { getServerSideApiServer } from '@common/api/utils';
-import { networkIndexState } from '@store/recoil/network';
-import { DEFAULT_NETWORK_INDEX, DEFAULT_TESTNET_INDEX } from '@common/constants';
 import { NetworkModeToast } from '@components/network-mode-toast';
+import { Modals } from '@components/modals';
 
 interface ExplorerAppProps extends AppProps {
-  networkMode: NetworkMode;
   apiServer: string;
+  networkMode: NetworkMode;
 }
 
 function ExplorerApp({
   Component,
-  networkMode,
   apiServer,
+  networkMode,
   pageProps: { isHome, fullWidth, ...props },
-}: ExplorerAppProps) {
+}: ExplorerAppProps): React.ReactElement {
   const router = useRouter();
-  const setNetworkIndex = useUpdateAtom(networkIndexState);
 
   useEffect(() => {
-    const chainMode = router.query.chain;
-    // Make sure the network list is in sync with the query param
-    if (chainMode && chainMode === 'testnet') {
-      setNetworkIndex(DEFAULT_TESTNET_INDEX);
-    } else {
-      setNetworkIndex(DEFAULT_NETWORK_INDEX);
-    }
-    toast(`You're viewing the ${chainMode || networkMode} Explorer`);
+    const chain = router.query.chain;
+    toast(`You're viewing the ${chain || networkMode} Explorer`);
   }, []);
 
   return (
@@ -48,6 +39,7 @@ function ExplorerApp({
       <AppConfig isHome={isHome} fullWidth={fullWidth}>
         <AtomDebug />
         <Component apiServer={apiServer} networkMode={networkMode} {...props} />
+        <Modals />
         <NetworkModeToast />
       </AppConfig>
     </>