+++ src/store/recoil/network.ts
@@ -3,10 +3,12 @@ import {
   DEFAULT_NETWORK_INDEX,
   DEFAULT_NETWORK_LIST,
   NETWORK_CURRENT_INDEX_COOKIE,
+  NETWORK_CUSTOM_LIST_COOKIE,
 } from '@common/constants';
 
 import { parseCookies, setCookie } from 'nookies';
 import { IS_SSR } from 'jotai-query-toolkit';
+import { NetworkMode, NetworkModes } from '@common/types/network';
 
 const atomWithCookie = <T>({ key, initialValue }: { key: string; initialValue: T }) => {
   const baseAtom = atom<T>(initialValue);
@@ -28,7 +30,7 @@ const atomWithCookie = <T>({ key, initialValue }: { key: string; initialValue: T
             maxAge: 30 * 24 * 60 * 60,
             path: '/',
           });
-          set(baseAtom, update);
+          void set(baseAtom, update);
         }
       }
     }
@@ -36,9 +38,12 @@ const atomWithCookie = <T>({ key, initialValue }: { key: string; initialValue: T
   return anAtom;
 };
 
-export const networkModeState = atom(null);
+export const networkModeState = atom<NetworkMode>(NetworkModes.Mainnet);
 
-export const customNetworksListState = atom<{ label: string; url: string }[]>([]);
+export const customNetworksListState = atomWithCookie<{ label: string; url: string }[]>({
+  key: NETWORK_CUSTOM_LIST_COOKIE,
+  initialValue: [],
+});
 
 export const networkListState = atom(get => {
   const customItems = get(customNetworksListState);