+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.14.8](https://github.com/blockstack/explorer/compare/v1.14.7...v1.14.8) (2021-09-20)
+
+
+### Bug Fixes
+
+* add network modal ([cf5821f](https://github.com/blockstack/explorer/commit/cf5821f1f91fc19a8919f4113e8eab9086fdc95d))
+* custom network list ([0bebb9a](https://github.com/blockstack/explorer/commit/0bebb9a9fb80e480bc134a57a7104568af5b5b8e))
+
+## [1.14.7](https://github.com/blockstack/explorer/compare/v1.14.6...v1.14.7) (2021-09-16)
+
+
+### Bug Fixes
+
+* regtest network mode ([5d43f21](https://github.com/blockstack/explorer/commit/5d43f21e2cb5c2589dde0d28d7c47a7853574ea4))
+
 ## [1.14.6](https://github.com/blockstack/explorer/compare/v1.14.5...v1.14.6) (2021-09-14)